Add Netlify config, update theme, add social media icons, and restore compatibility with Hugo 0.145.0 (#52)
This commit performs the following cleanups:
- Updates the theme submodule to be compatible with Hugo 0.144.0 (we
previously used Hugo 0.88.0 before I removed the binary in #51). I've
updated to the tip of the tree as of 08/03/2025 to use the latest
enhancements and ensure that we can build with new versions of Hugo,
i.e., 0.140.0 and later
- this also gets us Mastodon and LinkedIn social media icons that were
previously missing
- Dropped the "Categories" section from the navbar (it's currently empty
on https://blog.pyodide.org/categories/) for now, we can always
re-enable it later.
- Added a configuration file for Netlify to pin the Hugo version